Drastically improve search performance
